+++ examples/shardingsphere-jdbc-example/sharding-example/sharding-raw-jdbc-example/src/main/java/org/apache/shardingsphere/example/sharding/raw/jdbc/ShardingRawJavaConfigurationExample.java
@@ -19,11 +19,8 @@ package org.apache.shardingsphere.example.sharding.raw.jdbc;
 
 import org.apache.shardingsphere.example.core.api.ExampleExecuteTemplate;
 import org.apache.shardingsphere.example.core.api.service.ExampleService;
-import org.apache.shardingsphere.example.core.jdbc.repository.AddressRepositoryImpl;
-import org.apache.shardingsphere.example.core.jdbc.repository.OrderItemRepositoryImpl;
-import org.apache.shardingsphere.example.core.jdbc.repository.RangeOrderRepositoryImpl;
 import org.apache.shardingsphere.example.core.jdbc.service.OrderServiceImpl;
-import org.apache.shardingsphere.example.sharding.raw.jdbc.factory.RangeDataSourceFactory;
+import org.apache.shardingsphere.example.sharding.raw.jdbc.factory.DataSourceFactory;
 import org.apache.shardingsphere.example.type.ShardingType;
 
 import javax.sql.DataSource;
@@ -32,7 +29,7 @@ import java.sql.SQLException;
 /*
  * Please make sure primary replica data replication sync on MySQL is running correctly. Otherwise this example will query empty data from replica.
  */
-public final class JavaRangeConfigurationExampleMain {
+public final class ShardingRawJavaConfigurationExample {
 
     private static ShardingType shardingType = ShardingType.SHARDING_DATABASES;
 //    private static ShardingType shardingType = ShardingType.SHARDING_TABLES;
@@ -41,11 +38,11 @@ public final class JavaRangeConfigurationExampleMain {
 //    private static ShardingType shardingType = ShardingType.SHARDING_READWRITE_SPLITTING;
     
     public static void main(final String[] args) throws SQLException {
-        DataSource dataSource = RangeDataSourceFactory.newInstance(shardingType);
+        DataSource dataSource = DataSourceFactory.newInstance(shardingType);
         ExampleExecuteTemplate.run(getExampleService(dataSource));
     }
     
     private static ExampleService getExampleService(final DataSource dataSource) {
-        return new OrderServiceImpl(new RangeOrderRepositoryImpl(dataSource), new OrderItemRepositoryImpl(dataSource), new AddressRepositoryImpl(dataSource));
+        return new OrderServiceImpl(dataSource);
     }
 }
