+++ examples/shardingsphere-jdbc-example/sharding-example/sharding-spring-boot-jpa-example/src/main/resources/application-sharding-databases-tables.properties
@@ -17,14 +17,14 @@
 
 spring.shardingsphere.datasource.names=ds_0,ds_1
 
-spring.shardingsphere.datasource.ds_0.jdbc-url=jdbc:mysql://localhost:3306/demo_ds_0?serverTimezone=UTC&useSSL=false
+spring.shardingsphere.datasource.ds_0.jdbc-url=jdbc:mysql://localhost:3306/demo_ds_0?serverTimezone=UTC&useSSL=false&useUnicode=true&characterEncoding=UTF-8
 spring.shardingsphere.datasource.ds_0.type=com.zaxxer.hikari.HikariDataSource
 spring.shardingsphere.datasource.ds_0.driver-class-name=com.mysql.jdbc.Driver
 spring.shardingsphere.datasource.ds_0.username=root
 spring.shardingsphere.datasource.ds_0.password=
 spring.shardingsphere.datasource.ds_0.max-active=16
 
-spring.shardingsphere.datasource.ds_1.jdbc-url=jdbc:mysql://localhost:3306/demo_ds_1?serverTimezone=UTC&useSSL=false
+spring.shardingsphere.datasource.ds_1.jdbc-url=jdbc:mysql://localhost:3306/demo_ds_1?serverTimezone=UTC&useSSL=false&useUnicode=true&characterEncoding=UTF-8
 spring.shardingsphere.datasource.ds_1.type=com.zaxxer.hikari.HikariDataSource
 spring.shardingsphere.datasource.ds_1.driver-class-name=com.mysql.jdbc.Driver
 spring.shardingsphere.datasource.ds_1.username=root
