+++ examples/shardingsphere-jdbc-example/sharding-example/sharding-spring-boot-jpa-example/src/main/resources/application-sharding-readwrite-splitting.properties
@@ -17,42 +17,42 @@
 
 spring.shardingsphere.datasource.names=write_ds_0,write_ds_1,write_ds_0_read_0,write_ds_0_read_1,write_ds_1_read_0,write_ds_1_read_1
 
-spring.shardingsphere.datasource.write_ds_0.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_0?serverTimezone=UTC&useSSL=false
+spring.shardingsphere.datasource.write_ds_0.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_0?serverTimezone=UTC&useSSL=false&useUnicode=true&characterEncoding=UTF-8
 spring.shardingsphere.datasource.write_ds_0.type=com.zaxxer.hikari.HikariDataSource
 spring.shardingsphere.datasource.write_ds_0.driver-class-name=com.mysql.jdbc.Driver
 spring.shardingsphere.datasource.write_ds_0.username=root
 spring.shardingsphere.datasource.write_ds_0.password=
 spring.shardingsphere.datasource.write_ds_0.max-active=16
 
-spring.shardingsphere.datasource.write_ds_0_read_0.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_0_read_0?serverTimezone=UTC&useSSL=false
+spring.shardingsphere.datasource.write_ds_0_read_0.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_0_read_0?serverTimezone=UTC&useSSL=false&useUnicode=true&characterEncoding=UTF-8
 spring.shardingsphere.datasource.write_ds_0_read_0.type=com.zaxxer.hikari.HikariDataSource
 spring.shardingsphere.datasource.write_ds_0_read_0.driver-class-name=com.mysql.jdbc.Driver
 spring.shardingsphere.datasource.write_ds_0_read_0.username=root
 spring.shardingsphere.datasource.write_ds_0_read_0.password=
 spring.shardingsphere.datasource.write_ds_0_read_0.max-active=16
 
-spring.shardingsphere.datasource.write_ds_0_read_1.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_0_read_1?serverTimezone=UTC&useSSL=false
+spring.shardingsphere.datasource.write_ds_0_read_1.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_0_read_1?serverTimezone=UTC&useSSL=false&useUnicode=true&characterEncoding=UTF-8
 spring.shardingsphere.datasource.write_ds_0_read_1.type=com.zaxxer.hikari.HikariDataSource
 spring.shardingsphere.datasource.write_ds_0_read_1.driver-class-name=com.mysql.jdbc.Driver
 spring.shardingsphere.datasource.write_ds_0_read_1.username=root
 spring.shardingsphere.datasource.write_ds_0_read_1.password=
 spring.shardingsphere.datasource.write_ds_0_read_1.max-active=16
 
-spring.shardingsphere.datasource.write_ds_1.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_1?serverTimezone=UTC&useSSL=false
+spring.shardingsphere.datasource.write_ds_1.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_1?serverTimezone=UTC&useSSL=false&useUnicode=true&characterEncoding=UTF-8
 spring.shardingsphere.datasource.write_ds_1.type=com.zaxxer.hikari.HikariDataSource
 spring.shardingsphere.datasource.write_ds_1.driver-class-name=com.mysql.jdbc.Driver
 spring.shardingsphere.datasource.write_ds_1.username=root
 spring.shardingsphere.datasource.write_ds_1.password=
 spring.shardingsphere.datasource.write_ds_1.max-active=16
 
-spring.shardingsphere.datasource.write_ds_1_read_0.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_1_read_0?serverTimezone=UTC&useSSL=false
+spring.shardingsphere.datasource.write_ds_1_read_0.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_1_read_0?serverTimezone=UTC&useSSL=false&useUnicode=true&characterEncoding=UTF-8
 spring.shardingsphere.datasource.write_ds_1_read_0.type=com.zaxxer.hikari.HikariDataSource
 spring.shardingsphere.datasource.write_ds_1_read_0.driver-class-name=com.mysql.jdbc.Driver
 spring.shardingsphere.datasource.write_ds_1_read_0.username=root
 spring.shardingsphere.datasource.write_ds_1_read_0.password=
 spring.shardingsphere.datasource.write_ds_1_read_0.max-active=16
 
-spring.shardingsphere.datasource.write_ds_1_read_1.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_1_read_1?serverTimezone=UTC&useSSL=false
+spring.shardingsphere.datasource.write_ds_1_read_1.jdbc-url=jdbc:mysql://localhost:3306/demo_write_ds_1_read_1?serverTimezone=UTC&useSSL=false&useUnicode=true&characterEncoding=UTF-8
 spring.shardingsphere.datasource.write_ds_1_read_1.type=com.zaxxer.hikari.HikariDataSource
 spring.shardingsphere.datasource.write_ds_1_read_1.driver-class-name=com.mysql.jdbc.Driver
 spring.shardingsphere.datasource.write_ds_1_read_1.username=root
