+++ examples/shardingsphere-jdbc-example/sharding-example/sharding-spring-namespace-jpa-example/src/main/java/org/apache/shardingsphere/example/sharding/spring/namespace/jpa/ShardingSpringNamespaceJpaExample.java
@@ -15,26 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.example.sharding.spring.boot.jpa;
+package org.apache.shardingsphere.example.sharding.spring.namespace.jpa;
 
 import org.apache.shardingsphere.example.core.api.ExampleExecuteTemplate;
 import org.apache.shardingsphere.example.core.api.service.ExampleService;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.boot.autoconfigure.transaction.jta.JtaAutoConfiguration;
 import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import java.sql.SQLException;
 
-@ComponentScan("org.apache.shardingsphere.example.core.jpa")
-@EntityScan(basePackages = "org.apache.shardingsphere.example.core.jpa.entity")
-@SpringBootApplication(exclude = JtaAutoConfiguration.class)
-public class ExampleMain {
+public final class ShardingSpringNamespaceJpaExample {
+    
+    private static final String CONFIG_FILE = "META-INF/application-sharding-databases.xml";
+//    private static final String CONFIG_FILE = "META-INF/application-sharding-tables.xml";
+//    private static final String CONFIG_FILE = "META-INF/application-sharding-databases-tables.xml";
+//    private static final String CONFIG_FILE = "META-INF/application-readwrite-splitting.xml";
+//    private static final String CONFIG_FILE = "META-INF/application-sharding-readwrite-splitting.xml";
     
     public static void main(final String[] args) throws SQLException {
-        try (ConfigurableApplicationContext applicationContext = SpringApplication.run(ExampleMain.class, args)) {
+        try (ConfigurableApplicationContext applicationContext = new ClassPathXmlApplicationContext(CONFIG_FILE)) {
             ExampleExecuteTemplate.run(applicationContext.getBean(ExampleService.class));
         }
     }
